chore: append slices and maps while merging a container request
